renamed to ForecastDayCard.tsx to forecast_day_card.tsx for frontend file naming convention
