+++ frontend/src/components/forecast/forecast_day_card.tsx
